Split into multiple files and refactor
