+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "node-databox",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "A Nodejs library for interfacing with Databox APIs",
   "main": "databox.js",
   "scripts": {
